@@ -1,14 +1,15 @@
-import 'package:dropdown_banner/dropdown_banner.dart';
-import 'package:flutter/material.dart';
-import 'package:flutter/cupertino.dart';
-import 'package:skoller/tools.dart';
-import 'package:skoller/screens/auth/phone_verification_view.dart';
-import 'package:url_launcher/url_launcher.dart';
 import 'package:dart_notification_center/dart_notification_center.dart';
+import 'package:dropdown_banner/dropdown_banner.dart';
+import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
+import 'package:skoller/screens/auth/phone_verification_view.dart';
+import 'package:skoller/tools.dart';
+import 'package:url_launcher/url_launcher.dart';
+
 import 'sign_in.dart';
 
 class SignUp extends StatefulWidget {
-  SignUp({Key key}) : super(key: key);
+  SignUp({Key? key}) : super(key: key);
 
   @override
   _SignUpState createState() => _SignUpState();
@@ -17,12 +18,10 @@ class SignUp extends StatefulWidget {
 class _SignUpState extends State<SignUp> {
   final firstNameController = TextEditingController();
   final lastNameController = TextEditingController();
-  final emailController = TextEditingController();
   final phoneController = TextEditingController();
 
   final firstNameFocus = FocusNode();
   final lastNameFocus = FocusNode();
-  final emailFocus = FocusNode();
   final phoneFocus = FocusNode();
 
   bool validState = false;
@@ -32,11 +31,9 @@ class _SignUpState extends State<SignUp> {
     [
       firstNameController,
       lastNameController,
-      emailController,
       phoneController,
       firstNameFocus,
       lastNameFocus,
-      emailFocus,
       phoneFocus,
     ].forEach((obj) => (obj as dynamic).dispose());
 
@@ -46,13 +43,10 @@ class _SignUpState extends State<SignUp> {
   void verifyState(String _) {
     final firstName = firstNameController.text.trim();
     final lastName = lastNameController.text.trim();
-    final email = emailController.text.trim();
     final phone = phoneController.text.trim();
 
-    final isValid = firstName.length > 0 &&
-        lastName.length > 0 &&
-        email.contains(RegExp(r'(.+?)@(.+?)\.(.+?)')) &&
-        phone.length == 14;
+    final isValid =
+        firstName.length > 0 && lastName.length > 0 && phone.length == 14;
 
     if (isValid != validState) {
       setState(() => validState = isValid);
@@ -62,7 +56,6 @@ class _SignUpState extends State<SignUp> {
   void tappedSignUp(TapUpDetails details) {
     final firstName = firstNameController.text.trim();
     final lastName = lastNameController.text.trim();
-    final email = emailController.text.trim();
     final phone = phoneController.text.trim();
 
     final loadingScreen = SKLoadingScreen.fadeIn(context);
@@ -71,7 +64,6 @@ class _SignUpState extends State<SignUp> {
       nameFirst: firstName,
       nameLast: lastName,
       phone: phone.replaceAll(RegExp(r'[\(\) \-]+'), ''),
-      email: email,
     ).then((response) async {
       loadingScreen.fadeOut();
 
@@ -105,9 +97,9 @@ class _SignUpState extends State<SignUp> {
         }
       } else {
         setState(() => validState = false);
-        
+
         String message = [422, 401].contains(response.status)
-            ? 'A user already exists with that phone number or email. Try logging in!'
+            ? 'A user already exists with that phone number. Try logging in!'
             : 'Failed to create account. Try again later, or visit skoller.co';
 
         DropdownBanner.showBanner(
@@ -119,7 +111,7 @@ class _SignUpState extends State<SignUp> {
     });
   }
 
-  double dist;
+  double? dist;
 
   @override
   Widget build(BuildContext context) => Scaffold(
@@ -134,11 +126,10 @@ class _SignUpState extends State<SignUp> {
                 onVerticalDragCancel: () => dist = null,
                 onVerticalDragEnd: (details) => dist = null,
                 onVerticalDragUpdate: (details) {
-                  if (dist != null && details.localPosition.dy - dist > 30) {
+                  if (dist != null && details.localPosition.dy - dist! > 30) {
                     [
                       firstNameFocus,
                       lastNameFocus,
-                      emailFocus,
                       phoneFocus,
                     ].forEach((node) => node.hasFocus ? node.unfocus() : null);
 
@@ -215,8 +206,10 @@ class _SignUpState extends State<SignUp> {
                                     textCapitalization:
                                         TextCapitalization.words,
                                     textInputAction: TextInputAction.next,
-                                    onSubmitted: (_) =>
-                                        firstNameFocus.nextFocus(),
+                                    onSubmitted: (_) {
+                                      lastNameFocus.requestFocus();
+                                      // firstNameFocus.nextFocus();
+                                    },
                                     autofocus: true,
                                     onChanged: verifyState,
                                     focusNode: firstNameFocus,
@@ -261,8 +254,10 @@ class _SignUpState extends State<SignUp> {
                                     textCapitalization:
                                         TextCapitalization.words,
                                     textInputAction: TextInputAction.next,
-                                    onSubmitted: (_) =>
-                                        lastNameFocus.nextFocus(),
+                                    onSubmitted: (_) {
+                                      phoneFocus.requestFocus();
+                                      // return lastNameFocus.nextFocus();
+                                    },
                                     onChanged: verifyState,
                                     focusNode: lastNameFocus,
                                   ),
@@ -272,50 +267,6 @@ class _SignUpState extends State<SignUp> {
                           ),
                         ),
                       ],
-                    ),
-                    GestureDetector(
-                      onTapUp: (details) => emailFocus.requestFocus(),
-                      child: Container(
-                        margin:
-                            EdgeInsets.symmetric(vertical: 6, horizontal: 24),
-                        padding:
-                            EdgeInsets.symmetric(vertical: 4, horizontal: 8),
-                        decoration: BoxDecoration(
-                          color: Colors.white,
-                          borderRadius: BorderRadius.circular(5),
-                          border: Border.all(color: SKColors.border_gray),
-                          boxShadow: UIAssets.boxShadow,
-                        ),
-                        child: Column(
-                          crossAxisAlignment: CrossAxisAlignment.start,
-                          mainAxisSize: MainAxisSize.min,
-                          children: <Widget>[
-                            Text(
-                              'Email',
-                              style: TextStyle(
-                                  color: SKColors.skoller_blue,
-                                  fontSize: 12,
-                                  fontWeight: FontWeight.normal),
-                            ),
-                            CupertinoTextField(
-                              padding: EdgeInsets.all(1),
-                              controller: emailController,
-                              style: TextStyle(
-                                  fontSize: 15, color: SKColors.dark_gray),
-                              placeholder: 'School email recommended',
-                              placeholderStyle: TextStyle(
-                                  fontSize: 14,
-                                  color: SKColors.text_light_gray),
-                              decoration: BoxDecoration(border: null),
-                              textCapitalization: TextCapitalization.none,
-                              textInputAction: TextInputAction.next,
-                              onSubmitted: (_) => emailFocus.nextFocus(),
-                              onChanged: verifyState,
-                              focusNode: emailFocus,
-                            ),
-                          ],
-                        ),
-                      ),
                     ),
                     GestureDetector(
                       onTapUp: (details) => phoneFocus.requestFocus(),
@@ -379,7 +330,7 @@ class _SignUpState extends State<SignUp> {
                               TextSpan(
                                   text: 'User Agreement',
                                   style:
-                                      TextStyle(color: SKColors.skoller_blue))
+                                      TextStyle(color: SKColors.skoller_blue1))
                             ],
                           ),
                           textAlign: TextAlign.center,
@@ -395,7 +346,7 @@ class _SignUpState extends State<SignUp> {
                             behavior: HitTestBehavior.opaque,
                             child: Container(
                               alignment: Alignment.center,
-                              color: SKColors.skoller_blue,
+                              color: SKColors.skoller_blue1,
                               padding: EdgeInsets.symmetric(vertical: 12),
                               child: Text(
                                 'Sign up',
@@ -426,7 +377,7 @@ class _SignUpState extends State<SignUp> {
                                   child: Text(
                                     ' Log In',
                                     style: TextStyle(
-                                        color: SKColors.skoller_blue,
+                                        color: SKColors.skoller_blue1,
                                         fontWeight: FontWeight.bold),
                                   ),
                                 ),
