@@ -1,10 +1,10 @@
-import 'package:flutter/material.dart';
 import 'package:flutter/cupertino.dart';
-import 'package:skoller/tools.dart';
+import 'package:flutter/material.dart';
 import 'package:share/share.dart';
+import 'package:skoller/tools.dart';
 
 class ProfileLinkSharingView extends StatefulWidget {
-  final int startingClassId;
+  final int? startingClassId;
 
   ProfileLinkSharingView([this.startingClassId]);
 
@@ -17,7 +17,7 @@ enum _SharingType { classmates, anyone }
 class _ProfileLinkSharingState extends State<ProfileLinkSharingView> {
   var sharingType = _SharingType.classmates;
 
-  StudentClass studentClass;
+  StudentClass? studentClass;
 
   @override
   void initState() {
@@ -31,8 +31,8 @@ class _ProfileLinkSharingState extends State<ProfileLinkSharingView> {
 
   void tappedShare(_) {
     final shareMessage = sharingType == _SharingType.anyone
-        ? 'Check out this new app that\'s helping me keep up with school... it\'s like the Waze of the classroom!\n\n${SKUser.current.student.enrollmentLink ?? 'https://itunes.apple.com/us/app/skoller/id1314782490?mt=8'}'
-        : studentClass.shareMessage;
+        ? 'Check out this new app that\'s helping me keep up with school... it\'s like the Waze of the classroom!\n\n${SKUser.current?.student.enrollmentLink ?? 'https://itunes.apple.com/us/app/skoller/id1314782490?mt=8'}'
+        : studentClass!.shareMessage;
 
     Share.share(shareMessage);
   }
@@ -42,7 +42,7 @@ class _ProfileLinkSharingState extends State<ProfileLinkSharingView> {
     final classes = StudentClass.currentClasses.values.toList();
     final linkStripper = (String link) => link.split('//')[1];
 
-    final raiseEffort = SKUser.current.student.raiseEffort;
+    final raiseEffort = SKUser.current?.student.raiseEffort;
 
     return SafeArea(
       child: Column(
@@ -115,7 +115,7 @@ class _ProfileLinkSharingState extends State<ProfileLinkSharingView> {
                                           Text(
                                             'with ${sharingType == _SharingType.anyone ? 'anyone' : 'classmates'}',
                                             style: TextStyle(
-                                                color: SKColors.skoller_blue,
+                                                color: SKColors.skoller_blue1,
                                                 fontWeight: FontWeight.w500),
                                           ),
                                           Padding(
@@ -134,31 +134,12 @@ class _ProfileLinkSharingState extends State<ProfileLinkSharingView> {
                             SizedBox(width: 32),
                           ],
                         ),
-                        Padding(
-                          padding: EdgeInsets.symmetric(
-                              horizontal: 24, vertical: 24),
-                          child: raiseEffort?.orgId == null
-                              ? Text.rich(
-                                  TextSpan(
-                                    text: 'Get ',
-                                    children: [
-                                      TextSpan(
-                                        text: '5 classmates',
-                                        style: TextStyle(
-                                            fontWeight: FontWeight.bold),
-                                      ),
-                                      TextSpan(
-                                          text:
-                                              ' on Skoller and earn a \$10 Gift Card!')
-                                    ],
-                                    style: TextStyle(
-                                      fontWeight: FontWeight.w300,
-                                      fontSize: 16,
-                                    ),
-                                  ),
-                                  textAlign: TextAlign.center,
-                                )
-                              : Text.rich(
+                        raiseEffort?.orgId == null
+                            ? Container()
+                            : Padding(
+                                padding: EdgeInsets.symmetric(
+                                    horizontal: 24, vertical: 24),
+                                child: Text.rich(
                                   TextSpan(
                                     text: '',
                                     children: [
@@ -178,14 +159,14 @@ class _ProfileLinkSharingState extends State<ProfileLinkSharingView> {
                                   ),
                                   textAlign: TextAlign.center,
                                 ),
-                        ),
-                        Padding(
-                          padding: EdgeInsets.symmetric(horizontal: 24),
-                          child: AspectRatio(
-                            aspectRatio: 1,
-                            child: buildShareImage(),
-                          ),
-                        ),
+                              ),
+                        // Padding(
+                        //   padding: EdgeInsets.symmetric(horizontal: 24),
+                        //   child: AspectRatio(
+                        //     aspectRatio: 1,
+                        //     child: buildShareImage(),
+                        //   ),
+                        // ),
                         if (raiseEffort?.orgId != null)
                           Padding(
                             padding: EdgeInsets.symmetric(vertical: 8),
@@ -194,12 +175,12 @@ class _ProfileLinkSharingState extends State<ProfileLinkSharingView> {
                                 text: 'You\'ve raised ',
                                 children: [
                                   TextSpan(
-                                      text: '\$${raiseEffort.personalSignups}',
+                                      text: '\$${raiseEffort?.personalSignups}',
                                       style: TextStyle(
                                           fontWeight: FontWeight.bold)),
                                   TextSpan(
                                       text:
-                                          ' for the ${raiseEffort.orgName == 'AOII' ? 'Arthritis Foundation' : 'ASA Foundation'}!')
+                                          ' for the ${raiseEffort?.orgName == 'AOII' ? 'Arthritis Foundation' : 'ASA Foundation'}!')
                                 ],
                                 style: TextStyle(fontWeight: FontWeight.normal),
                               ),
@@ -221,8 +202,8 @@ class _ProfileLinkSharingState extends State<ProfileLinkSharingView> {
                                     subtitle:
                                         'Let your classmates use your class specific link to enroll',
                                     items: classes
-                                        .map(
-                                            (studentClass) => studentClass.name)
+                                        .map((studentClass) =>
+                                            studentClass.name!)
                                         .toList(),
                                     onSelect: (index) {
                                       setState(
@@ -237,14 +218,14 @@ class _ProfileLinkSharingState extends State<ProfileLinkSharingView> {
                                 decoration: BoxDecoration(
                                   borderRadius: BorderRadius.circular(5),
                                   border:
-                                      Border.all(color: SKColors.skoller_blue),
+                                      Border.all(color: SKColors.skoller_blue1),
                                 ),
                                 child: Text(
                                   studentClass == null
                                       ? 'Select class...'
-                                      : studentClass.name,
+                                      : studentClass!.name!,
                                   style:
-                                      TextStyle(color: SKColors.skoller_blue),
+                                      TextStyle(color: SKColors.skoller_blue1),
                                 ),
                               ),
                             ),
@@ -261,8 +242,8 @@ class _ProfileLinkSharingState extends State<ProfileLinkSharingView> {
                               child: Container(
                                 decoration: BoxDecoration(
                                   color: sharingType == _SharingType.anyone
-                                      ? SKColors.skoller_blue
-                                      : studentClass.getColor(),
+                                      ? SKColors.skoller_blue1
+                                      : studentClass!.getColor(),
                                   borderRadius: BorderRadius.circular(5),
                                   boxShadow: UIAssets.boxShadow,
                                 ),
@@ -279,9 +260,10 @@ class _ProfileLinkSharingState extends State<ProfileLinkSharingView> {
                                     Text(
                                       linkStripper(
                                         sharingType == _SharingType.anyone
-                                            ? SKUser
-                                                .current.student.enrollmentLink
-                                            : studentClass.enrollmentLink,
+                                            ? SKUser.current?.student
+                                                    .enrollmentLink ??
+                                                ''
+                                            : studentClass!.enrollmentLink,
                                       ),
                                       style: TextStyle(color: Colors.white),
                                     ),
@@ -290,7 +272,7 @@ class _ProfileLinkSharingState extends State<ProfileLinkSharingView> {
                               ),
                             ),
                           ),
-                        Padding(
+                        /* Padding(
                           padding: EdgeInsets.fromLTRB(8, 8, 8, 0),
                           child: Text(
                             raiseEffort?.orgId == null
@@ -302,7 +284,7 @@ class _ProfileLinkSharingState extends State<ProfileLinkSharingView> {
                                 fontWeight: FontWeight.normal,
                                 fontSize: 13),
                           ),
-                        ),
+                        ),*/
                       ],
                     ),
                   ),
@@ -316,7 +298,7 @@ class _ProfileLinkSharingState extends State<ProfileLinkSharingView> {
   }
 
   Image buildShareImage() {
-    final orgName = SKUser.current.student.raiseEffort?.orgName;
+    final orgName = SKUser.current?.student.raiseEffort?.orgName;
 
     if (orgName == 'AOII')
       return Image.asset(ImageNames.shareImages.aoii_share_image);
