@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
 import 'package:skoller/screens/main_app/tutorial/calendar_tutorial_view.dart';
 import 'package:skoller/tools.dart';
+
 import './menu/school_search_view.dart';
 
 class PrimarySchoolModal extends StatefulWidget {
@@ -10,69 +11,71 @@ class PrimarySchoolModal extends StatefulWidget {
 }
 
 class _PrimarySchoolState extends State<PrimarySchoolModal> {
-  List<School> eligibleSchools;
+  late List<School>? eligibleSchools = null;
 
-  int selectedSchoolId;
-  Period selectedPeriod;
+  int? selectedSchoolId;
+  Period? selectedPeriod;
 
   bool showingGreeting = true;
-  bool loading;
+  bool loading = false;
 
   @override
   void initState() {
     super.initState();
 
-    if (SKUser.current.student.primarySchool != null) {
-      eligibleSchools = [SKUser.current.student.primarySchool];
-      selectedPeriod = eligibleSchools.first.getBestCurrentPeriod();
+    if (SKUser.current?.student.primarySchool != null) {
+      eligibleSchools = [SKUser.current!.student.primarySchool!];
+      selectedPeriod = eligibleSchools?.first.getBestCurrentPeriod();
       showingGreeting = false;
-    } else {
-      loading = true;
-
-      SKUser.current.checkEmailDomain().then((response) async {
-        if (response.wasSuccessful()) {
-          final List<School> obj = response.obj;
-
-          if (obj.length == 1) {
-            selectedSchoolId = obj.first.id;
-            selectedPeriod = obj.first.getBestCurrentPeriod();
-            await SKUser.current.update(primarySchool: obj.first);
-          } else if (obj.length == 0 && !showingGreeting) tappedSearch(null);
-
-          eligibleSchools = obj;
-        }
-
-        setState(() {
-          loading = false;
-        });
-      }).catchError((_) => setState(() => loading = false));
+      // } else {
+      //   loading = true;
+      //
+      //   SKUser.current?.checkEmailDomain().then((response) async {
+      //     if (response.wasSuccessful()) {
+      //       final List<School> obj = response.obj;
+      //
+      //       if (obj.length == 1) {
+      //         selectedSchoolId = obj.first.id;
+      //         selectedPeriod = obj.first.getBestCurrentPeriod();
+      //         await SKUser.current?.update(primarySchool: obj.first);
+      //       } else if (obj.length == 0 && !showingGreeting) tappedSearch(null);
+      //
+      //       eligibleSchools = obj;
+      //     }
+      //
+      //     setState(() {
+      //       loading = false;
+      //     });
+      //   }).catchError((_) => setState(() => loading = false));
     }
   }
 
-  void tappedSearch(TapUpDetails details) async {
+  void tappedSearch(TapUpDetails? details) async {
     await Navigator.push(
       context,
       SKNavFadeUpRoute(builder: (context) => SchoolSearchView()),
     );
 
-    if (SKUser.current.student.primarySchool != null)
+    if (SKUser.current?.student.primarySchool != null)
       setState(() {
-        eligibleSchools = [SKUser.current.student.primarySchool];
-        selectedPeriod = eligibleSchools.first.getBestCurrentPeriod();
+        eligibleSchools = [SKUser.current!.student.primarySchool!];
+        selectedPeriod = eligibleSchools!.first.getBestCurrentPeriod();
       });
   }
 
   void tappedPeriodSelect(TapUpDetails detail) {
     final now = DateTime.now();
 
-    final eligiblePeriods = eligibleSchools.first.periods == null
+    final eligiblePeriods = (eligibleSchools?.first.periods == null
         ? null
-        : eligibleSchools.first.periods.toList()
-      ..removeWhere((period) => now.isAfter(period.endDate))
+        : eligibleSchools?.first.periods!.toList())!;
+    final List<Period> remainingPeriods = eligiblePeriods
+      ..removeWhere((period) => now.isAfter(period.endDate!));
+    final List<Period> sortedPeriods = remainingPeriods
       ..sort(
         (period1, period2) {
           return period2.startDate != null
-              ? (period1.startDate?.compareTo(period2.startDate) ?? -1)
+              ? (period1.startDate?.compareTo(period2.startDate!) ?? -1)
               : 1;
         },
       );
@@ -84,33 +87,33 @@ class _PrimarySchoolState extends State<PrimarySchoolModal> {
         title: 'Active term',
         subtitle: 'Which term are you using Skoller for right now?',
         onSelect: (index) =>
-            setState(() => selectedPeriod = eligiblePeriods[index]),
-        items: eligiblePeriods.toList().map((p) => p.name).toList(),
+            setState(() => selectedPeriod = sortedPeriods[index]),
+        items: sortedPeriods.toList().map((p) => p.name).toList(),
       ),
     );
   }
 
-  void tappedSchoolSelect(TapUpDetails details) async {
+  void tappedSchoolSelect(TapUpDetails? details) async {
     if (selectedPeriod != null) {
       final loader = SKLoadingScreen.fadeIn(context);
-      await SKUser.current.update(primaryPeriod: selectedPeriod);
+      await SKUser.current?.update(primaryPeriod: selectedPeriod!);
       loader.fadeOut();
 
       Navigator.pop(context);
     } else if (selectedSchoolId != null) {
-      final school = eligibleSchools.firstWhere(
+      final school = eligibleSchools?.firstWhere(
           (school) => school.id == selectedSchoolId,
-          orElse: () => eligibleSchools.first);
+          orElse: () => eligibleSchools!.first);
 
-      await SKUser.current.update(primarySchool: school);
+      await SKUser.current?.update(primarySchool: school);
 
-      if (SKUser.current.student.primaryPeriod != null &&
-          SKUser.current.student.primarySchool != null)
+      if (SKUser.current?.student.primaryPeriod != null &&
+          SKUser.current?.student.primarySchool != null)
         Navigator.pop(context);
       else
         setState(() {
-          eligibleSchools = [SKUser.current.student.primarySchool];
-          selectedPeriod = eligibleSchools.first.getBestCurrentPeriod();
+          eligibleSchools = [SKUser.current!.student.primarySchool!];
+          selectedPeriod = eligibleSchools!.first.getBestCurrentPeriod();
         });
     }
   }
@@ -137,17 +140,17 @@ class _PrimarySchoolState extends State<PrimarySchoolModal> {
                   crossAxisAlignment: CrossAxisAlignment.stretch,
                   mainAxisSize: MainAxisSize.min,
                   children: [
-                    if (eligibleSchools == null)
-                      Row(
-                        mainAxisAlignment: MainAxisAlignment.center,
-                        children: <Widget>[
-                          SizedBox(
-                              width: 40,
-                              height: 40,
-                              child: CircularProgressIndicator()),
-                        ],
-                      ),
-                    if ((eligibleSchools?.length ?? -1) == 0) ...buildSearch(),
+                    // if (eligibleSchools == null)
+                    //   Row(
+                    //     mainAxisAlignment: MainAxisAlignment.center,
+                    //     children: <Widget>[
+                    //       SizedBox(
+                    //           width: 40,
+                    //           height: 40,
+                    //           child: CircularProgressIndicator()),
+                    //     ],
+                    //   ),
+                    if (eligibleSchools == null) ...buildSearch(),
                     if ((eligibleSchools?.length ?? -1) == 1) ...buildSingle(),
                     if ((eligibleSchools?.length ?? -1) > 1) ...buildMultiple(),
                   ]),
@@ -328,12 +331,12 @@ class _PrimarySchoolState extends State<PrimarySchoolModal> {
                       child: SizedBox(
                         width: 32,
                         height: 32,
-                        child: CircularProgressIndicator(),
+                        child: Center(child: CircularProgressIndicator()),
                       ),
                     )
                   : GestureDetector(
                       onTapUp: (details) {
-                        if (eligibleSchools.length == 0) tappedSearch(null);
+                        if (eligibleSchools?.length == 0) tappedSearch(null);
                         setState(() => showingGreeting = false);
                       },
                       child: Container(
@@ -343,7 +346,7 @@ class _PrimarySchoolState extends State<PrimarySchoolModal> {
                             EdgeInsets.symmetric(horizontal: 24, vertical: 8),
                         decoration: BoxDecoration(
                             borderRadius: BorderRadius.circular(5),
-                            color: SKColors.skoller_blue,
+                            color: SKColors.skoller_blue1,
                             boxShadow: UIAssets.boxShadow),
                         child: Text(
                           'Get started',
@@ -436,14 +439,14 @@ class _PrimarySchoolState extends State<PrimarySchoolModal> {
   }
 
   List<Widget> buildSingle() {
-    final school = eligibleSchools[0];
+    final school = eligibleSchools![0];
     final period = selectedPeriod;
 
     final formatter = DateFormat('MMMM');
     final start =
-        period?.startDate == null ? null : formatter.format(period.startDate);
+        period?.startDate == null ? null : formatter.format(period!.startDate!);
     final end =
-        period?.endDate == null ? null : formatter.format(period.endDate);
+        period?.endDate == null ? null : formatter.format(period!.endDate!);
 
     return [
       SammiSpeechBubble(
@@ -562,7 +565,7 @@ class _PrimarySchoolState extends State<PrimarySchoolModal> {
           margin: EdgeInsets.fromLTRB(24, 24, 24, 16),
           padding: EdgeInsets.symmetric(vertical: 8),
           decoration: BoxDecoration(
-            color: SKColors.skoller_blue,
+            color: SKColors.skoller_blue1,
             borderRadius: BorderRadius.circular(5),
             boxShadow: UIAssets.boxShadow,
           ),
@@ -605,7 +608,7 @@ class _PrimarySchoolState extends State<PrimarySchoolModal> {
                   crossAxisAlignment: CrossAxisAlignment.stretch,
                   mainAxisSize: MainAxisSize.min,
                   children: [
-                    ...eligibleSchools
+                    ...eligibleSchools!
                         .map(
                           (school) => GestureDetector(
                             onTapUp: (details) =>
