@@ -1,12 +1,12 @@
-import 'package:flutter/material.dart';
 import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
 import 'package:share/share.dart';
 import 'package:skoller/tools.dart';
 
 class RewardsView extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
-    final raiseEffort = SKUser.current.student.raiseEffort;
+    final raiseEffort = SKUser.current?.student.raiseEffort;
     return SKNavView(
       title: 'My Rewards',
       leftBtn: Image.asset(ImageNames.navArrowImages.down),
@@ -14,7 +14,7 @@ class RewardsView extends StatelessWidget {
         Flexible(
             child: raiseEffort?.orgId == null
                 ? createPoints()
-                : createRaiseEffort(raiseEffort)),
+                : createRaiseEffort(raiseEffort!)),
       ],
     );
   }
@@ -78,7 +78,7 @@ class RewardsView extends StatelessWidget {
                       Padding(
                         padding: EdgeInsets.only(left: 8),
                         child: Text(
-                          SKUser.current.student.primarySchool.name,
+                          SKUser.current?.student.primarySchool?.name ?? '',
                           style: TextStyle(
                               fontWeight: FontWeight.normal,
                               fontSize: 13,
@@ -155,11 +155,11 @@ class RewardsView extends StatelessWidget {
             border: Border.all(color: SKColors.border_gray),
             boxShadow: UIAssets.boxShadow),
         child: Column(children: [
-          Text(
+          /*Text(
             'You earn rewards when other \nstudents sign up using your links.',
             style: TextStyle(fontWeight: FontWeight.w300, fontSize: 14),
             textAlign: TextAlign.center,
-          ),
+          ),*/
           Padding(
             padding: EdgeInsets.all(16),
             child: Image.asset(ImageNames.shareImages.amazon_card_image),
@@ -169,7 +169,7 @@ class RewardsView extends StatelessWidget {
             child: Text.rich(
               TextSpan(
                 text:
-                    '${SKUser.current.student.raiseEffort?.personalSignups ?? 0} students\n',
+                    '${SKUser.current?.student.raiseEffort?.personalSignups ?? 0} students\n',
                 children: [
                   TextSpan(
                       text: 'have signed up using your links',
@@ -184,14 +184,14 @@ class RewardsView extends StatelessWidget {
           GestureDetector(
             onTapUp: (details) {
               Share.share(
-                  'Check out this new app that\'s helping me keep up with school... it\'s like the Waze of the classroom!\n\n${SKUser.current.student.enrollmentLink ?? 'https://itunes.apple.com/us/app/skoller/id1314782490?mt=8'}');
+                  'Check out this new app that\'s helping me keep up with school... it\'s like the Waze of the classroom!\n\n${SKUser.current?.student.enrollmentLink ?? 'https://itunes.apple.com/us/app/skoller/id1314782490?mt=8'}');
             },
             child: Container(
               alignment: Alignment.center,
               padding: EdgeInsets.symmetric(vertical: 8),
               margin: EdgeInsets.symmetric(horizontal: 16, vertical: 12),
               decoration: BoxDecoration(
-                color: SKColors.skoller_blue,
+                color: SKColors.skoller_blue1,
                 borderRadius: BorderRadius.circular(5),
               ),
               child: Text(
@@ -200,14 +200,14 @@ class RewardsView extends StatelessWidget {
               ),
             ),
           ),
-          Text(
+          /*Text(
             'A Skoller employee will email you when you qualify!',
             textAlign: TextAlign.center,
             style: TextStyle(
                 fontSize: 13,
                 fontWeight: FontWeight.normal,
                 color: SKColors.text_light_gray),
-          ),
+          ),*/
         ]),
       );
 }
