@@ -1,17 +1,229 @@
-import 'package:skoller/screens/main_app/menu/profile_link_sharing_view.dart';
 import 'package:dart_notification_center/dart_notification_center.dart';
-import 'package:skoller/screens/main_app/menu/manage_classes_view.dart';
-import 'package:skoller/screens/main_app/menu/rewards_view.dart';
-import 'package:skoller/screens/main_app/menu/reminders_view.dart';
-import 'package:skoller/screens/main_app/menu/profile_view.dart';
-import 'package:skoller/screens/main_app/tutorial/tutorial.dart';
-import 'package:url_launcher/url_launcher.dart';
-import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
+import 'package:skoller/screens/main_app/menu/manage_classes_view.dart';
+import 'package:skoller/screens/main_app/menu/profile_link_sharing_view.dart';
+import 'package:skoller/screens/main_app/menu/profile_view.dart';
+import 'package:skoller/screens/main_app/menu/rewards_view.dart';
+import 'package:skoller/screens/main_app/premium/account_setting_dialog_view.dart';
+import 'package:skoller/screens/main_app/premium/premium_packages_view.dart';
+import 'package:skoller/screens/main_app/tutorial/tutorial.dart';
 import 'package:skoller/tools.dart';
+import 'package:url_launcher/url_launcher.dart';
+
+import 'premium/already_premium_view.dart';
 
 class MenuView extends StatelessWidget {
-  final List<Widget> menuOptions = [
+  List<Widget> menuOptions(context) {
+    return [
+      [
+        {
+          'name': Text(
+            'Profile',
+            style: TextStyle(fontWeight: FontWeight.normal, fontSize: 16),
+          ),
+          'builder': () => ProfileView(),
+          'image': Image.asset(ImageNames.peopleImages.person_blue)
+        },
+        {
+          'name': Text(
+            'Account Settings',
+            style: TextStyle(fontWeight: FontWeight.normal, fontSize: 16),
+          ),
+          'builder1': () => AlreadyPremiumView(),
+          'image': Image.asset(ImageNames.peopleImages.person_blue)
+        }
+      ],
+      [
+        {
+          'name': Text(
+            'Share Skoller',
+            style: TextStyle(fontWeight: FontWeight.normal, fontSize: 16),
+          ),
+          'builder': () => ProfileLinkSharingView(),
+          'image': Image.asset(ImageNames.peopleImages.people_blue)
+        },
+        {
+          'name': Text(
+            'My rewards',
+            style: TextStyle(fontWeight: FontWeight.normal, fontSize: 16),
+          ),
+          'builder': () => RewardsView(),
+          'image': Image.asset(ImageNames.menuImages.points)
+        },
+      ],
+      [
+        {
+          'name': Text(
+            'Reminders',
+            style: TextStyle(fontWeight: FontWeight.normal, fontSize: 16),
+          ),
+          // 'builder': () => AlreadyPremiumView(),
+          'image': Image.asset(ImageNames.menuImages.reminders)
+        },
+        {
+          'name': Text(
+            'Manage classes',
+            style: TextStyle(fontWeight: FontWeight.normal, fontSize: 16),
+          ),
+          'builder': () => ManageClassesView(),
+          'image': Image.asset('image_assets/tab_bar_assets/classes_blue.png')
+        },
+      ],
+      [
+        {
+          'name': Text(
+            'Send us feedback',
+            style: TextStyle(fontWeight: FontWeight.normal, fontSize: 16),
+          ),
+          'action': () async {
+            final url = 'mailto:support@skoller.co?subject=Feedback';
+
+            if (await canLaunch(url)) launch(url);
+          },
+        },
+        {
+          'name': Text(
+            'Tutorial',
+            style: TextStyle(fontWeight: FontWeight.normal, fontSize: 16),
+          ),
+          'builder': () => TutorialTab(
+                (newContext) => Navigator.of(newContext).pop(),
+                'Dismiss',
+              ),
+        },
+        // {
+        //   'name': Text(
+        //     'Payment',
+        //     style: TextStyle(fontWeight: FontWeight.normal, fontSize: 16),
+        //   ),
+        //   'builder': () => PremiumPackagesView()
+        // },
+      ],
+      [
+        {
+          'name': Text(
+            'Privacy Policy',
+            style: TextStyle(fontWeight: FontWeight.normal, fontSize: 16),
+          ),
+          'action': () async {
+            final url = 'https://skoller.co/privacypolicy';
+            if (await canLaunch(url)) launch(url);
+          },
+        },
+        {
+          'name': Text(
+            'Terms of Use',
+            style: TextStyle(fontWeight: FontWeight.normal, fontSize: 16),
+          ),
+          'action': () async {
+            final url = 'https://skoller.co/useragreement';
+            if (await canLaunch(url)) launch(url);
+          },
+        },
+      ]
+    ]
+        .map((group) => Container(
+              decoration: BoxDecoration(
+                color: Colors.white,
+                border: Border(bottom: BorderSide(color: SKColors.border_gray)),
+              ),
+              margin: EdgeInsets.symmetric(vertical: 8),
+              child: Column(
+                crossAxisAlignment: CrossAxisAlignment.stretch,
+                children: group
+                    .map(
+                      (row) => Container(
+                        height: 44,
+                        decoration: BoxDecoration(
+                            border: Border(
+                                top: BorderSide(color: SKColors.border_gray))),
+                        padding: EdgeInsets.symmetric(horizontal: 12),
+                        child: GestureDetector(
+                          onTap: () {},
+                          behavior: HitTestBehavior.opaque,
+                          onTapUp: (details) {
+                            if (row.containsKey('builder1')) {
+                              showDialog(
+                                context: context,
+                                barrierDismissible: false,
+                                builder: (_) => AccountSettingsDialogView(),
+                              );
+                              // loadData(context);
+                            }
+                            if (row.containsKey('builder')) {
+                              String channel;
+                              Widget view = (row['builder'] as Function)();
+
+                              if (view is ProfileLinkSharingView)
+                                channel = NotificationChannels
+                                    .presentModalViewOverTabBar;
+                              else
+                                channel =
+                                    NotificationChannels.presentViewOverTabBar;
+
+                              DartNotificationCenter.post(
+                                channel: channel,
+                                options: view,
+                              );
+                            } else if (row.containsKey('action'))
+                              (row['action'] as VoidCallback)();
+                          },
+                          child: Row(
+                            children: [
+                              ...row.containsKey('image')
+                                  ? [
+                                      Padding(
+                                        padding: EdgeInsets.only(
+                                            right: 8, bottom: 1),
+                                        child: SizedBox(
+                                            width: 20,
+                                            height: 20,
+                                            child: row['image'] as Widget),
+                                      ),
+                                    ]
+                                  : [],
+                              Expanded(child: row['name'] as Widget),
+                            ],
+                          ),
+                        ),
+                      ),
+                    )
+                    .toList(),
+              ),
+            ))
+        .toList();
+  }
+
+  loadData(context) {
+    // if (Subscriptions.isLifetimeSubscription || Subscriptions.isLifetimeTrial) {
+    //   showDialog(
+    //     context: context,
+    //     builder: (_) => AlreadyPremiumView(),
+    //   );
+    // }
+
+    if (Subscriptions.isSubscriptionActive) {
+      showDialog(
+        context: context,
+        builder: (_) => PremiumPackagesView(true),
+        // builder: (_) => AlreadyPremiumView(),
+      );
+    } else {
+      if (Subscriptions.isTrial) {
+        showDialog(
+          context: context,
+          builder: (_) => AlreadyPremiumView(),
+        );
+      } else {
+        showDialog(
+          context: context,
+          builder: (_) => PremiumPackagesView(true),
+        );
+      }
+    }
+  }
+
+  /*  List<Widget> menuOptions = [
     [
       {
         'name': Text(
@@ -19,6 +231,14 @@ class MenuView extends StatelessWidget {
           style: TextStyle(fontWeight: FontWeight.normal, fontSize: 16),
         ),
         'builder': () => ProfileView(),
+        'image': Image.asset(ImageNames.peopleImages.person_blue)
+      },
+      {
+        'name': Text(
+          'Account Settings',
+          style: TextStyle(fontWeight: FontWeight.normal, fontSize: 16),
+        ),
+        'builder1': () => AlreadyPremiumView(),
         'image': Image.asset(ImageNames.peopleImages.person_blue)
       }
     ],
@@ -46,7 +266,7 @@ class MenuView extends StatelessWidget {
           'Reminders',
           style: TextStyle(fontWeight: FontWeight.normal, fontSize: 16),
         ),
-        'builder': () => RemindersView(),
+        // 'builder': () => AlreadyPremiumView(),
         'image': Image.asset(ImageNames.menuImages.reminders)
       },
       {
@@ -80,6 +300,13 @@ class MenuView extends StatelessWidget {
               'Dismiss',
             ),
       },
+      // {
+      //   'name': Text(
+      //     'Payment',
+      //     style: TextStyle(fontWeight: FontWeight.normal, fontSize: 16),
+      //   ),
+      //   'builder': () => PremiumPackagesView()
+      // },
     ]
   ]
       .map((group) => Container(
@@ -99,8 +326,17 @@ class MenuView extends StatelessWidget {
                               top: BorderSide(color: SKColors.border_gray))),
                       padding: EdgeInsets.symmetric(horizontal: 12),
                       child: GestureDetector(
+                        onTap: () {},
                         behavior: HitTestBehavior.opaque,
                         onTapUp: (details) {
+                          if (row.containsKey('builder1')) {
+                            Navigator.pop(context);
+                            showDialog(
+                                context: context,
+                                builder: (_) {
+                                  return AlreadyPremiumView();
+                                });
+                          }
                           if (row.containsKey('builder')) {
                             String channel;
                             Widget view = (row['builder'] as Function)();
@@ -129,11 +365,11 @@ class MenuView extends StatelessWidget {
                                       child: SizedBox(
                                           width: 20,
                                           height: 20,
-                                          child: row['image']),
+                                          child: row['image'] as Widget),
                                     ),
                                   ]
                                 : [],
-                            Expanded(child: row['name']),
+                            Expanded(child: row['name'] as Widget),
                           ],
                         ),
                       ),
@@ -142,7 +378,159 @@ class MenuView extends StatelessWidget {
                   .toList(),
             ),
           ))
-      .toList();
+      .toList();*/
+
+  // final List<Widget> menuOptions = [
+  //   [
+  //     {
+  //       'name': Text(
+  //         'Profile',
+  //         style: TextStyle(fontWeight: FontWeight.normal, fontSize: 16),
+  //       ),
+  //       'builder': () => ProfileView(),
+  //       'image': Image.asset(ImageNames.peopleImages.person_blue)
+  //     },
+  //     {
+  //       'name': Text(
+  //         'Account Settings',
+  //         style: TextStyle(fontWeight: FontWeight.normal, fontSize: 16),
+  //       ),
+  //       'builder': (context) => showDialog(
+  //           context: context,
+  //           builder: (_) {
+  //             return AlreadyPremiumView();
+  //           }),
+  //       'image': Image.asset(ImageNames.peopleImages.person_blue)
+  //     }
+  //   ],
+  //   [
+  //     {
+  //       'name': Text(
+  //         'Share Skoller',
+  //         style: TextStyle(fontWeight: FontWeight.normal, fontSize: 16),
+  //       ),
+  //       'builder': () => ProfileLinkSharingView(),
+  //       'image': Image.asset(ImageNames.peopleImages.people_blue)
+  //     },
+  //     {
+  //       'name': Text(
+  //         'My rewards',
+  //         style: TextStyle(fontWeight: FontWeight.normal, fontSize: 16),
+  //       ),
+  //       'builder': () => RewardsView(),
+  //       'image': Image.asset(ImageNames.menuImages.points)
+  //     },
+  //   ],
+  //   [
+  //     {
+  //       'name': Text(
+  //         'Reminders',
+  //         style: TextStyle(fontWeight: FontWeight.normal, fontSize: 16),
+  //       ),
+  //       // 'builder': () => AlreadyPremiumView(),
+  //       'image': Image.asset(ImageNames.menuImages.reminders)
+  //     },
+  //     {
+  //       'name': Text(
+  //         'Manage classes',
+  //         style: TextStyle(fontWeight: FontWeight.normal, fontSize: 16),
+  //       ),
+  //       'builder': () => ManageClassesView(),
+  //       'image': Image.asset('image_assets/tab_bar_assets/classes_blue.png')
+  //     },
+  //   ],
+  //   [
+  //     {
+  //       'name': Text(
+  //         'Send us feedback',
+  //         style: TextStyle(fontWeight: FontWeight.normal, fontSize: 16),
+  //       ),
+  //       'action': () async {
+  //         final url = 'mailto:support@skoller.co?subject=Feedback';
+  //
+  //         if (await canLaunch(url)) launch(url);
+  //       },
+  //     },
+  //     {
+  //       'name': Text(
+  //         'Tutorial',
+  //         style: TextStyle(fontWeight: FontWeight.normal, fontSize: 16),
+  //       ),
+  //       'builder': () => TutorialTab(
+  //             (newContext) => Navigator.of(newContext).pop(),
+  //             'Dismiss',
+  //           ),
+  //     },
+  //     // {
+  //     //   'name': Text(
+  //     //     'Payment',
+  //     //     style: TextStyle(fontWeight: FontWeight.normal, fontSize: 16),
+  //     //   ),
+  //     //   'builder': () => PremiumPackagesView()
+  //     // },
+  //   ]
+  // ]
+  //     .map((group) => Container(
+  //           decoration: BoxDecoration(
+  //             color: Colors.white,
+  //             border: Border(bottom: BorderSide(color: SKColors.border_gray)),
+  //           ),
+  //           margin: EdgeInsets.symmetric(vertical: 8),
+  //           child: Column(
+  //             crossAxisAlignment: CrossAxisAlignment.stretch,
+  //             children: group
+  //                 .map(
+  //                   (row) => Container(
+  //                     height: 44,
+  //                     decoration: BoxDecoration(
+  //                         border: Border(
+  //                             top: BorderSide(color: SKColors.border_gray))),
+  //                     padding: EdgeInsets.symmetric(horizontal: 12),
+  //                     child: GestureDetector(
+  //                       behavior: HitTestBehavior.opaque,
+  //                       onTapUp: (details) {
+  //                         if (row.containsKey('builder')) {
+  //                           String channel;
+  //                           Widget view = (row['builder'] as Function)();
+  //
+  //                           if (view is ProfileLinkSharingView)
+  //                             channel = NotificationChannels
+  //                                 .presentModalViewOverTabBar;
+  //                           else
+  //                             channel =
+  //                                 NotificationChannels.presentViewOverTabBar;
+  //
+  //                           DartNotificationCenter.post(
+  //                             channel: channel,
+  //                             options: view,
+  //                           );
+  //                         } else if (row.containsKey('action'))
+  //                           (row['action'] as VoidCallback)();
+  //                       },
+  //                       child: Row(
+  //                         children: [
+  //                           ...row.containsKey('image')
+  //                               ? [
+  //                                   Padding(
+  //                                     padding:
+  //                                         EdgeInsets.only(right: 8, bottom: 1),
+  //                                     child: SizedBox(
+  //                                         width: 20,
+  //                                         height: 20,
+  //                                         child: row['image'] as Widget),
+  //                                   ),
+  //                                 ]
+  //                               : [],
+  //                           Expanded(child: row['name'] as Widget),
+  //                         ],
+  //                       ),
+  //                     ),
+  //                   ),
+  //                 )
+  //                 .toList(),
+  //           ),
+  //         ))
+  //     .toList();
 
   @override
   Widget build(BuildContext context) {
@@ -170,20 +558,21 @@ class MenuView extends StatelessWidget {
                         decoration: BoxDecoration(
                           color: SKColors.light_gray,
                           shape: BoxShape.circle,
-                          image: SKUser.current.avatarUrl == null
+                          image: SKUser.current?.avatarUrl == null
                               ? null
                               : DecorationImage(
                                   fit: BoxFit.fill,
-                                  image: NetworkImage(SKUser.current.avatarUrl),
+                                  image:
+                                      NetworkImage(SKUser.current!.avatarUrl!),
                                 ),
                         ),
                         margin: EdgeInsets.only(left: 12),
                         height: 44,
                         width: 44,
-                        child: SKUser.current.avatarUrl == null
+                        child: SKUser.current?.avatarUrl == null
                             ? Text(
-                                SKUser.current.student.nameFirst[0] +
-                                    SKUser.current.student.nameLast[0],
+                                SKUser.current!.student.nameFirst![0] +
+                                    SKUser.current!.student.nameLast![0],
                                 style: TextStyle(
                                   color: Colors.white,
                                   fontSize: 21,
@@ -197,7 +586,7 @@ class MenuView extends StatelessWidget {
                           padding:
                               EdgeInsets.symmetric(vertical: 24, horizontal: 8),
                           child: Text(
-                            '${SKUser.current.student.nameFirst} ${SKUser.current.student.nameLast}',
+                            '${SKUser.current?.student.nameFirst} ${SKUser.current?.student.nameLast}',
                             style: TextStyle(fontSize: 18),
                           ),
                         ),
@@ -207,7 +596,7 @@ class MenuView extends StatelessWidget {
                   Expanded(
                     child: ListView(
                       physics: ClampingScrollPhysics(),
-                      children: menuOptions,
+                      children: menuOptions(context),
                     ),
                   ),
                   Text(
